lib: obstacles: move class List to its own file
