+++ examples/lidar_obstacles/lidar_obstacles.cpp
@@ -12,6 +12,7 @@
 #include "common_defs.h"
 #include "lidar_utils.hpp"
 #include "obstacles/obstacles.hpp"
+#include "obstacles/List.hpp"
 #include "cogip_defs/Coords.hpp"
 #include "cogip_defs/Pose.hpp"
