+++ examples/lidar_obstacles/main.cpp
@@ -5,6 +5,7 @@
 // Project includes
 #include "shell_menu/shell_menu.hpp"
 #include "obstacles/obstacles.hpp"
+#include "obstacles/List.hpp"
 #include "tracefd/tracefd.hpp"
 
 // Application includes
