+++ lib/obstacles/obstacles_private.hpp
@@ -1,0 +1,14 @@
+#include "obstacles/List.hpp"
+
+// System includes
+#include <set>
+
+namespace cogip {
+
+namespace obstacles {
+
+extern std::set<List const *> all_obstacles;
+
+} // namespace obstacles
+
+} // namespace cogip
